requirements.txt
blueliv-python-sdk==2.3.0
certifi==2017.7.27.1
chardet==3.0.4
fake-useragent==0.1.7
idna==2.5
pkg-resources==0.0.0
python-dateutil==2.6.1
requests==2.5.1
six==1.10.0
urllib3==1.22
